+++ sentry-android-core/src/main/java/io/sentry/android/core/internal/gestures/ViewUtils.java
@@ -7,7 +7,6 @@ import android.view.ViewGroup;
 import io.sentry.android.core.SentryAndroidOptions;
 import io.sentry.internal.gestures.GestureTargetLocator;
 import io.sentry.internal.gestures.UiElement;
-import io.sentry.util.Objects;
 import java.util.LinkedList;
 import java.util.Queue;
 import org.jetbrains.annotations.ApiStatus;
@@ -16,6 +15,32 @@ import org.jetbrains.annotations.Nullable;
 
 @ApiStatus.Internal
 public final class ViewUtils {
+
+  private static final int[] coordinates = new int[2];
+
+  /**
+   * Verifies if the given touch coordinates are within the bounds of the given view.
+   *
+   * @param view the view to check if the touch coordinates are within its bounds
+   * @param x - the x coordinate of a {@link MotionEvent}
+   * @param y - the y coordinate of {@link MotionEvent}
+   * @return true if the touch coordinates are within the bounds of the view, false otherwise
+   */
+  private static boolean touchWithinBounds(
+      final @Nullable View view, final float x, final float y) {
+    if (view == null) {
+      return false;
+    }
+
+    view.getLocationOnScreen(coordinates);
+    int vx = coordinates[0];
+    int vy = coordinates[1];
+
+    int w = view.getWidth();
+    int h = view.getHeight();
+
+    return !(x < vx || x > vx + w || y < vy || y > vy + h);
+  }
 
   /**
    * Finds a target view, that has been selected/clicked by the given coordinates x and y and the
@@ -40,7 +65,12 @@ public final class ViewUtils {
 
     @Nullable UiElement target = null;
     while (queue.size() > 0) {
-      final View view = Objects.requireNonNull(queue.poll(), "view is required");
+      final View view = queue.poll();
+
+      if (!touchWithinBounds(view, x, y)) {
+        // if the touch is not hitting the view, skip traversal of its children
+        continue;
+      }
 
       if (view instanceof ViewGroup) {
         final ViewGroup viewGroup = (ViewGroup) view;
@@ -54,7 +84,7 @@ public final class ViewUtils {
         if (newTarget != null) {
           if (targetType == UiElement.Type.CLICKABLE) {
             target = newTarget;
-          } else {
+          } else if (targetType == UiElement.Type.SCROLLABLE) {
             return newTarget;
           }
         }