+++ sentry/src/test/java/io/sentry/instrumentation/file/SentryFileOutputStreamTest.kt
@@ -14,6 +14,8 @@ import org.junit.rules.TemporaryFolder
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.whenever
 import java.io.File
+import java.io.FileOutputStream
+import java.io.IOException
 import java.util.concurrent.atomic.AtomicBoolean
 import kotlin.concurrent.thread
 import kotlin.test.Test
@@ -25,6 +27,7 @@ import kotlin.test.assertTrue
 class SentryFileOutputStreamTest {
     class Fixture {
         val hub = mock<IHub>()
+        val options = SentryOptions()
         lateinit var sentryTracer: SentryTracer
 
         internal fun getSut(
@@ -32,17 +35,32 @@ class SentryFileOutputStreamTest {
             activeTransaction: Boolean = true,
             append: Boolean = false
         ): SentryFileOutputStream {
-            whenever(hub.options).thenReturn(
-                SentryOptions().apply {
-                    mainThreadChecker = MainThreadChecker.getInstance()
-                    addInAppInclude("org.junit")
-                }
-            )
+            options.run {
+                mainThreadChecker = MainThreadChecker.getInstance()
+                addInAppInclude("org.junit")
+            }
+            whenever(hub.options).thenReturn(options)
             sentryTracer = SentryTracer(TransactionContext("name", "op"), hub)
             if (activeTransaction) {
                 whenever(hub.span).thenReturn(sentryTracer)
             }
             return SentryFileOutputStream(tmpFile, append, hub)
+        }
+
+        internal fun getSut(
+            tmpFile: File? = null,
+            delegate: FileOutputStream,
+            tracesSampleRate: Double? = 1.0
+        ): FileOutputStream {
+            options.tracesSampleRate = tracesSampleRate
+            whenever(hub.options).thenReturn(options)
+            sentryTracer = SentryTracer(TransactionContext("name", "op"), hub)
+            whenever(hub.span).thenReturn(sentryTracer)
+            return SentryFileOutputStream.Factory.create(
+                delegate,
+                tmpFile,
+                hub
+            )
         }
     }
 
@@ -156,5 +174,26 @@ class SentryFileOutputStreamTest {
         val fileIOSpan = fixture.sentryTracer.children.first()
         assertEquals(false, fileIOSpan.data[SpanDataConvention.BLOCKED_MAIN_THREAD_KEY])
         assertNull(fileIOSpan.data[SpanDataConvention.CALL_STACK_KEY])
+    }
+
+    @Test
+    fun `when tracing is disabled does not instrument the stream`() {
+        val file = tmpFile
+        val delegate = ThrowingFileOutputStream(file)
+        val stream = fixture.getSut(file, delegate = delegate, tracesSampleRate = null)
+
+        assertTrue { stream is ThrowingFileOutputStream }
+    }
+}
+
+class ThrowingFileOutputStream(file: File) : FileOutputStream(file) {
+    val throwable = IOException("Oops!")
+
+    override fun write(b: Int) {
+        throw throwable
+    }
+
+    override fun close() {
+        throw throwable
     }
 }